+++ packages/nextjs/components/Card/GigCard/types.ts
@@ -1,4 +1,4 @@
-import { Gig } from "~~/types/gig.types";
+import { Gig } from "~~/types/gig/gig.types";
 
 export interface GigCardProps extends React.HTMLAttributes<HTMLDivElement> {
   gig: Gig;
