+++ packages/nextjs/components/Card/JobPostingCard/types.ts
@@ -1,4 +1,4 @@
-import { JobPosting } from "~~/types/job.types";
+import { JobPosting } from "~~/types/job";
 
 export interface JobPostingCardProps extends React.HTMLAttributes<HTMLDivElement> {
   jobPosting?: JobPosting;
