+++ packages/nextjs/components/GigApplicationsList/GigApplicationsList.tsx
@@ -17,8 +17,8 @@ import {
   XCircleIcon,
 } from "@heroicons/react/24/outline";
 // Adjust the import to match the actual export from the module
-import { fetchApplicationsWithGigDetails } from "~~/services/graphql/fetchers/gig.service";
-import { Application, ApplicationsData, Gig, GigState } from "~~/types/gig.types";
+import { fetchApplicationsWithGigDetails } from "~~/services/graphql/fetchers/gig/gig.service";
+import { Application, ApplicationsData, Gig, GigState } from "~~/types/gig/gig.types";
 
 interface InfoIcons {
   title: string;
