+++ packages/nextjs/components/JobsList/JobsList.tsx
@@ -17,8 +17,8 @@ import {
   XCircleIcon,
 } from "@heroicons/react/24/outline";
 // Adjust the import to match the actual export from the module
-import { fetchMyJobs } from "~~/services/graphql/fetchers/job.service";
-import { Job, JobState, JobsData } from "~~/types/job.types";
+import { fetchMyJobs } from "~~/services/graphql/fetchers/job/job.service";
+import { Job, JobState, JobsData } from "~~/types/job/job.types";
 
 interface InfoIcons {
   title: string;
